rename codecov.js to coverage-targets
